Add: delete confirmation added

+++ src/app/feature/product/view-products/view-products.component.ts
@@ -76,15 +76,17 @@ export class ViewProductsComponent {
   }
 
   deleteProduct(productId: string) {
-    this.apiService.deleteProduct(productId).subscribe({
-      next: (response) => {
-        this.snackBar.open("deleted successfully", 'Close', { duration: 5000, panelClass: 'success-snackbar' });
-        this.getProductsByCompanyId(this.companyId);
-      },
-      error: (error) => {
-        this.snackBar.open(error?.error?.message ?? 'An error occurred while deleting', 'Close', { duration: 5000, panelClass: 'error-snackbar' });
-      },
-    })
+    if (confirm(`Are you sure you want to delete`)) {
+      this.apiService.deleteProduct(productId).subscribe({
+        next: (response) => {
+          this.snackBar.open("deleted successfully", 'Close', { duration: 5000, panelClass: 'success-snackbar' });
+          this.getProductsByCompanyId(this.companyId);
+        },
+        error: (error) => {
+          this.snackBar.open(error?.error?.message ?? 'An error occurred while deleting', 'Close', { duration: 5000, panelClass: 'error-snackbar' });
+        },
+      })
+    }
   }
   reloadList() {
     this.selectedProduct.set(this.getEmptyProductObject());
